backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openpyxl==3.1.2
networkx==3.2.1
scikit-learn==1.4.2
sentence-transformers==2.7.0
torch==2.2.2
chromadb==0.5.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.29
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic-settings==2.2.1
python-louvain==0.16
numpy==1.26.4
scipy==1.13.0
aiohttp==3.9.5
aiofiles==23.2.1
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
